test: organize test modules
