+++ .vscode/settings.json
@@ -5,9 +5,13 @@
     "python.analysis.inlayHints.functionReturnTypes": false,
     "python.analysis.inlayHints.variableTypes": true,
     "python.analysis.inlayHints.pytestParameters": false,
-    "python.analysis.enablePytestSupport": false,
+    "python.analysis.enablePytestSupport": true,
     "editor.formatOnSave": true,
     "python.terminal.activateEnvironment": true,
+    "python.defaultInterpreterPath": "${workspaceFolder}/.venv/bin/python",
+    "python.envFile": "${workspaceFolder}/.vscode/test.env",
+    "python.testing.cwd": "${workspaceFolder}",
+    "python.testing.pytestPath": "${workspaceFolder}/.venv/bin/pytest",
     "python.analysis.extraPaths": [
         "./src"
     ],
