+++ .vscode/test.env
@@ -1,0 +1,3 @@
+PYTHONPATH=src
+ISH_CONFIG_PATH=/tmp/ish-vscode-test
+ISH_METRICS_PORT=
