+++ tests/conftest.py
@@ -1,0 +1,5 @@
+import os
+
+
+os.environ.setdefault("ISH_CONFIG_PATH", "/tmp/ish-test-config")
+os.environ.setdefault("ISH_METRICS_PORT", "")
